+++ src/App.js
@@ -8,44 +8,44 @@ import customTheme from './theme/customTheme';
 // Components
 import Login from './components/Auth/Login';
 import Signup from './components/Auth/Signup';
-import ReviewForm from './components/Dashboard/ReviewForm';
+import ReviewForm from './components/Dashboard/Forms/ReviewForm';
 import PrivateRoute from './utils/PrivateRoute';
 import SingleReview from './components/Dashboard/SingleReview';
 import DashboardHome from './components/Dashboard/DashboardHome';
-import AddCompanyForm from './components/Dashboard/AddCompanyForm';
+import AddCompanyForm from './components/Dashboard/Forms/AddCompanyForm';
 // google analytics
 import ReactGA from 'react-ga';
 import { useLocation } from 'react-router-dom';
 
 function initializeAnalytics() {
-  return process.env.REACT_APP_env === 'production'
-    ? ReactGA.initialize('UA-159325981-1') &&
-        // ReactGA.ga('send', 'pageview', window.location.pathname)
-        ReactGA.pageview(window.location.pathname + window.location.search)
-    : null;
+	return process.env.REACT_APP_env === 'production'
+		? ReactGA.initialize('UA-159325981-1') &&
+				// ReactGA.ga('send', 'pageview', window.location.pathname)
+				ReactGA.pageview(window.location.pathname + window.location.search)
+		: null;
 }
 
 const App = () => {
-  const location = useLocation();
+	const location = useLocation();
 
-  useEffect(() => {
-    initializeAnalytics();
-  }, [location]);
+	useEffect(() => {
+		initializeAnalytics();
+	}, [location]);
 
-  return (
-    <ThemeProvider theme={customTheme}>
-      <div className='App'>
-        <Switch>
-          <Route exact path='/' component={Login} />
-          <Route path='/signup' component={Signup} />
-          <PrivateRoute exact path='/dashboard' component={DashboardHome} />
-          <PrivateRoute path='/dashboard/add-review' component={ReviewForm} />
-          <PrivateRoute path='/dashboard/:id' component={SingleReview} />
-          <PrivateRoute path='/add-company' component={AddCompanyForm} />
-        </Switch>
-      </div>
-    </ThemeProvider>
-  );
+	return (
+		<ThemeProvider theme={customTheme}>
+			<div className='App'>
+				<Switch>
+					<Route exact path='/' component={Login} />
+					<Route path='/signup' component={Signup} />
+					<PrivateRoute exact path='/dashboard' component={DashboardHome} />
+					<PrivateRoute path='/dashboard/add-review' component={ReviewForm} />
+					<PrivateRoute path='/dashboard/:id' component={SingleReview} />
+					<PrivateRoute path='/add-company' component={AddCompanyForm} />
+				</Switch>
+			</div>
+		</ThemeProvider>
+	);
 };
 
 export default App;
